+++ src/server/Cartera/CarteraRouteExternal.ts
@@ -1,6 +1,7 @@
 import * as express from "express";
 import { iNode } from "../iNode";
 import { CarteraServer } from "./CarteraServer";
+import { postSendReceiptsEmailHandler } from './SendReceiptEmail';
 
 const router = express.Router();
 
@@ -11,6 +12,14 @@ router.get("/rebuild-pending-payments-month", (req, res) => {
     response: res
   };
   server.rebuildPendingPaymentsForMonthHandler(node);
+});
+
+router.post("/send-receipts-email", (req, res) => {
+  let node: iNode = {
+    request: req,
+    response: res
+  };
+  postSendReceiptsEmailHandler(node);
 });
 
 export { router };